refactor: Update user_data_replace_on_change variable in main.tf
This commit updates the user_data_replace_on_change variable in the main.tf file. Instead of hardcoding it to true, it now uses the value provided in the user_data_replace_on_change variable defined in variables.tf. This change allows for more flexibility and customization when replacing user data on instance changes.